requirements.txt
aiofiles==25.1.0
aiogram==3.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2026.2.25
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.23.0
python-dotenv==1.0.1